Move serial ControlBits type to uefi-raw
